requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.9.0

# LangChain Ecosystem (0.3.x 최신 호환 버전)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-elasticsearch>=0.3.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0

# AI & Data
openai>=1.50.0
elasticsearch>=8.15.0
huggingface-hub>=0.26.0
sentence-transformers>=3.2.0
torch>=2.5.0
numpy<2.0.0
pandas>=2.2.0

# Tools & Utils
google-search-results>=2.4.2
requests>=2.32.0
tiktoken>=0.8.0
pypdf>=5.0.0
tqdm>=4.66.0
streamlit>=1.39.0